Made Matrix a managed type with a View
Added a function to RingBuffer that outputs whether the buffer is filled. This test is now used in DNNlayer to fix an issue when pipelining where later ranks in the pipeline were performed actions on uninitialized inputs until the pipeline is primed.
This is benign in the sense that it does not make the ultimate output incorrect, but led to SEGVs when using the new Matrix implementation because no allocations are performed for zero-sized matrices.

ManagedArray now throws an error when attempting to open a view on a zero-sized array.